Split the sqlite and postgres configurations
