+++ apps/acme-order/Db/OrderContext.cs
@@ -1,90 +1,18 @@
-﻿using System;
-using System.Collections.Generic;
-using acme_order.Models;
+﻿using acme_order.Models;
 using Microsoft.EntityFrameworkCore;
-using Newtonsoft.Json;
+using Microsoft.Extensions.Configuration;
 
 namespace acme_order.Db
 {
-    public class OrderContext : DbContext
+    public abstract class OrderContext : DbContext
     {
-        public OrderContext(DbContextOptions<OrderContext> options)
-            : base(options)
+        protected readonly IConfiguration Configuration;
+
+        protected OrderContext(IConfiguration configuration)
         {
+            Configuration = configuration;
         }
 
         public virtual DbSet<Order> Orders { get; set; }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-        }
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.HasPostgresExtension("uuid-ossp");
-            modelBuilder.Entity<Order>(entity =>
-            {
-                entity.ToTable("order");
-
-                entity.Property(e => e.Id)
-                    .HasColumnName("id")
-                    .HasColumnType("uuid")
-                    .HasDefaultValueSql("uuid_generate_v4()")
-                    .IsRequired();
-
-                entity.Property(e => e.Address)
-                    .HasColumnName("address")
-                    .HasColumnType("json")
-                    .HasConversion(
-                        v => JsonConvert.SerializeObject(v, new JsonSerializerSettings { NullValueHandling = NullValueHandling.Ignore }),
-                        v => JsonConvert.DeserializeObject<Address>(v, new JsonSerializerSettings { NullValueHandling = NullValueHandling.Ignore }));
-
-                entity.Property(e => e.Card)
-                    .HasColumnName("card")
-                    .HasColumnType("json")
-                    .HasConversion(
-                        v => JsonConvert.SerializeObject(v, new JsonSerializerSettings { NullValueHandling = NullValueHandling.Ignore }),
-                        v => JsonConvert.DeserializeObject<Card>(v, new JsonSerializerSettings { NullValueHandling = NullValueHandling.Ignore }));
-
-                entity.Property(e => e.Cart)
-                    .HasColumnName("cart")
-                    .HasColumnType("json")
-                    .HasConversion(
-                        v => JsonConvert.SerializeObject(v, new JsonSerializerSettings { NullValueHandling = NullValueHandling.Ignore }),
-                        v => JsonConvert.DeserializeObject<ICollection<Cart>>(v, new JsonSerializerSettings { NullValueHandling = NullValueHandling.Ignore }));
-
-                entity.Property(e => e.Date)
-                    .HasColumnName("date")
-                    .HasDefaultValue(DateTime.UtcNow);
-
-                entity.Property(e => e.Delivery)
-                    .HasColumnName("delivery")
-                    .HasMaxLength(1000);
-
-                entity.Property(e => e.Email)
-                    .HasColumnName("email")
-                    .HasMaxLength(1000);
-
-                entity.Property(e => e.Firstname)
-                    .HasColumnName("firstname")
-                    .HasMaxLength(1000);
-
-                entity.Property(e => e.Lastname)
-                    .HasColumnName("lastname")
-                    .HasMaxLength(1000);
-
-                entity.Property(e => e.Paid)
-                    .HasColumnName("paid")
-                    .HasMaxLength(1000);
-
-                entity.Property(e => e.Total)
-                    .HasColumnName("total")
-                    .HasMaxLength(1000);
-
-                entity.Property(e => e.UserId)
-                    .HasColumnName("user_id")
-                    .HasMaxLength(1000);
-            });
-        }
     }
 }
