+++ apps/acme-order/acme-order.csproj
@@ -26,4 +26,9 @@
     <PackageReference Include="Steeltoe.Management.EndpointCore" Version="3.1.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Migrations\Postgres" />
+    <Folder Include="Migrations\Sqlite" />
+  </ItemGroup>
+
 </Project>
