+++ apps/acme-order/appsettings.json
@@ -4,8 +4,9 @@
       "Name":"order-service"
     }
   },
+  "DatabaseProvider": "Sqlite",
   "ConnectionStrings": {
-    "OrderContext": "",
+    "OrderContext": "Data Source=sqlite.order.db",
     "KeyVaultUri": ""
   },
   "ApplicationInsights": {
